pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "arroyoGISAXS"
version = "0.0.1"
description = "A package to perform computations and suggestions during a time-resolved GISAXS experiment."
authors = [{ name = "Dylan McReynolds", email = "dmcreynolds@lbl.gov" }]
dependencies = [
    "arroyopy",
    "dynaconf",
    "python-dotenv",
    "pandas",
    "msgpack",
    "numpy",
    "Pillow",
    "pyzmq",
    "tqdm",
    "typer",
    "websockets",
    "zarr",
    "tiled >=0.1.0b14, <0.2",
    "redis"]

[project.optional-dependencies]
workflow-viz = [
    "joblib",
    "numpy<2.0.0",
    "pandas",
    "plotly-express",
    "pyFAI==2023.9.0",
    "python-dotenv",
    "pyarrow>=14.0.1",
    "requests==2.26.0",
    "diskcache==5.6.3",
    "scikit-learn==1.3.0"]

dev = [
    "fakeredis",
    "flake8",
    "pre-commit",
    "pytest-asyncio",
    "pytest-mock",
    "tiled[server] >=0.1.0b14, <0.2"
    ]

lse = [
    "torchvision==0.17.2",
    "joblib"
]

all = [
    "arroyopy",
    "dynaconf",
    "python-dotenv",
    "pandas",
    "msgpack",
    "numpy",
    "Pillow",
    "pyzmq",
    "tqdm",
    "typer",
    "websockets",
    "zarr",
    "tiled[client] >=0.1.0b14, <0.2",
    "redis",
    "joblib",
    "numpy<2.0.0",
    "pandas",
    "plotly-express",
    "pyFAI==2023.9.0",
    "python-dotenv",
    "pyarrow>=14.0.1",
    "requests==2.26.0",
    "diskcache==5.6.3",
    "scikit-learn==1.3.0",
    # "torchvision==0.17.2",
]

[tool.isort]
profile = "black"


[tool.pixi.project]
name = "arroyoGISAXS"
version = "0.0.1"
channels = ["conda-forge"]
platforms = ["linux-64"]

[tool.pixi.tasks]
pre-commit-install = "pre-commit install"
pre-commit = "pre-commit run --all-files"
start_ws = "python -m arroyogisaxs.websockets"


[tool.pixi.pypi-dependencies]


[tool.pixi.environments]
# # default = { solve-group = "default" }
dev = { features = ["dev"] }
# # notebook = { features = ["notebook"], solve-group = "default" }
# lse = { features = ["lse"]}
